requirements.txt
accelerate==0.33.0
aiodns==3.2.0
aiohappyeyeballs==2.3.5
aiohttp==3.10.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
boto3==1.34.162
botocore==1.34.162
Brotli==1.1.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.8
diffusers==0.30.0
dnspython==2.6.1
email_validator==2.2.0
fastapi==0.112.1
fastapi-cli==0.0.5
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.1
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.24.5
idna==3.7
importlib_metadata==8.2.0
inquirerpy==0.3.4
itsdangerous==2.2.0
Jinja2==3.1.4
jmespath==1.0.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
networkx>=3.2.1
numpy==1.26.4
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.20.5
nvidia-nvjitlink-cu12==12.6.20
nvidia-nvtx-cu12==12.1.105
orjson==3.10.7
packaging==24.1
paramiko==3.4.1
pfzy==0.3.4
pillow==10.4.0
prettytable==3.11.0
prompt_toolkit==3.0.47
protobuf==5.27.3
psutil==6.0.0
py-cpuinfo==9.0.0
pycares==4.4.0
pycparser==2.22
pydantic==2.8.2
pydantic-extra-types==2.9.0
pydantic-settings==2.4.0
pydantic_core==2.20.1
Pygments==2.18.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.2
regex==2024.7.24
requests==2.32.3
rich==13.7.1
runpod==1.7.0
s3transfer==0.10.2
safetensors==0.4.4
sentencepiece==0.2.0
setuptools==72.2.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.2
sympy==1.13.2
tokenizers==0.19.1
tomli==2.0.1
tomlkit==0.13.2
torch==2.4.0
tqdm==4.66.5
tqdm-loggable==0.2
transformers==4.44.0
triton==3.0.0
typer==0.12.3
typing_extensions==4.12.2
ujson==5.10.0
urllib3>=1.25.4
uvicorn==0.30.6
uvloop==0.20.0
watchdog==4.0.2
watchfiles==0.23.0
wcwidth==0.2.13
websockets==12.0
xformers==0.0.27.post2
yarl==1.9.4
zipp==3.20.0
